requirements.txt

streamlit
youtube-transcript-api
python-docx
